HUM-652 : rename / move replicationDevice
the replicationDevice as it was basically is swift disk layout specific. Now
that we are moving on to EC / Indexdb it is confusing to have it in the
Replicator because there is a whole lot of gnarly code (REPCON , old disk
layouts, etc) that is there just to support the old way of doing things.

The replicator will continue to be used so this makes looking at that file not
quite so bad.
