+++ objectserver/replicator.go
@@ -16,25 +16,17 @@
 package objectserver
 
 import (
-	"encoding/hex"
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"math/rand"
 	"net/http"
 	_ "net/http/pprof"
-	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/troubling/hummingbird/common"
 	"github.com/troubling/hummingbird/common/conf"
-	"github.com/troubling/hummingbird/common/fs"
-	"github.com/troubling/hummingbird/common/pickle"
 	"github.com/troubling/hummingbird/common/ring"
 	"github.com/troubling/hummingbird/common/srv"
 	"github.com/troubling/hummingbird/middleware"
@@ -62,75 +54,11 @@ type PriorityRepJob struct {
 	Policy     int          `json:"policy"`
 }
 
-type quarantineFileError struct {
-	msg string
-}
-
-func (q quarantineFileError) Error() string {
-	return q.msg
-}
-
 func deviceKeyId(dev string, policy int) string {
 	if policy == 0 {
 		return dev
 	}
 	return fmt.Sprintf("%s-%d", dev, policy)
-}
-
-func getFile(filePath string) (fp *os.File, xattrs []byte, size int64, err error) {
-	fp, err = os.Open(filePath)
-	if err != nil {
-		return nil, nil, 0, fmt.Errorf("unable to open file (%v): %s", err, filePath)
-	}
-	defer func() {
-		if err != nil {
-			fp.Close()
-		}
-	}()
-	finfo, err := fp.Stat()
-	if err != nil || !finfo.Mode().IsRegular() {
-		return nil, nil, 0, quarantineFileError{"not a regular file"}
-	}
-	rawxattr, err := common.SwiftObjectRawReadMetadata(fp.Fd())
-	if err != nil || len(rawxattr) == 0 {
-		return nil, nil, 0, quarantineFileError{"error reading xattrs"}
-	}
-
-	// Perform a mini-audit, since it's cheap and we can potentially avoid spreading bad data around.
-	v, err := pickle.PickleLoads(rawxattr)
-	if err != nil {
-		return nil, nil, 0, quarantineFileError{"error unpickling xattrs"}
-	}
-	metadata, ok := v.(map[interface{}]interface{})
-	if !ok {
-		return nil, nil, 0, quarantineFileError{"invalid metadata type"}
-	}
-	for key, value := range metadata {
-		if _, ok := key.(string); !ok {
-			return nil, nil, 0, quarantineFileError{"invalid key in metadata"}
-		}
-		if _, ok := value.(string); !ok {
-			return nil, nil, 0, quarantineFileError{"invalid value in metadata"}
-		}
-	}
-	switch filepath.Ext(filePath) {
-	case ".data":
-		for _, reqEntry := range []string{"Content-Length", "Content-Type", "name", "ETag", "X-Timestamp"} {
-			if _, ok := metadata[reqEntry]; !ok {
-				return nil, nil, 0, quarantineFileError{".data missing required metadata"}
-			}
-		}
-		if contentLength, err := strconv.ParseInt(metadata["Content-Length"].(string), 10, 64); err != nil || contentLength != finfo.Size() {
-			return nil, nil, 0, quarantineFileError{"invalid content-length"}
-		}
-	case ".ts":
-		for _, reqEntry := range []string{"name", "X-Timestamp"} {
-			if _, ok := metadata[reqEntry]; !ok {
-				return nil, nil, 0, quarantineFileError{".ts missing required metadata"}
-			}
-		}
-	}
-	return fp, rawxattr, finfo.Size(), nil
 }
 
 type DeviceStats struct {
@@ -149,600 +77,11 @@ type DeviceStats struct {
 	PriorityRepsDone   int64
 }
 
-type replJob struct {
-	partition string
-	nodes     []*ring.Device
-	headers   map[string]string
-}
-
-type replicationDevice struct { //TODO: move all replicationDevice stuff into swiftobjeng.go
-	// If you have a better way to make struct methods that are overridable for tests, please call my house.
-	i interface {
-		beginReplication(dev *ring.Device, partition string, hashes bool, rChan chan beginReplicationResponse, headers map[string]string)
-		listObjFiles(objChan chan string, cancel chan struct{}, partdir string, needSuffix func(string) bool)
-		syncFile(objFile string, dst []*syncFileArg, handoff bool) (syncs int, insync int, err error)
-		replicateUsingHashes(rjob replJob, moreNodes ring.MoreNodes)
-		replicateAll(rjob replJob, isHandoff bool)
-		cleanTemp()
-		listPartitions() ([]string, []string, error)
-		replicatePartition(partition string)
-	}
-	r      *Replicator
-	dev    *ring.Device
-	policy int
-	cancel chan struct{}
-	priRep chan PriorityRepJob
-}
-
 type statUpdate struct {
 	service   string
 	deviceKey string
 	stat      string
 	value     int64
-}
-
-func (rd *replicationDevice) UpdateStat(stat string, amount int64) {
-	rd.r.updateStat <- statUpdate{"object-replicator", rd.Key(), stat, amount}
-}
-
-type beginReplicationResponse struct {
-	dev    *ring.Device
-	conn   RepConn
-	hashes map[string]string
-	err    error
-}
-
-func (rd *replicationDevice) listObjFiles(objChan chan string, cancel chan struct{}, partdir string, needSuffix func(string) bool) {
-	defer close(objChan)
-	suffixDirs, err := filepath.Glob(filepath.Join(partdir, "[a-f0-9][a-f0-9][a-f0-9]"))
-	if err != nil {
-		rd.r.logger.Error("[listObjFiles]", zap.Error(err))
-		return
-	}
-	if len(suffixDirs) == 0 {
-		os.Remove(filepath.Join(partdir, ".lock"))
-		os.Remove(filepath.Join(partdir, "hashes.pkl"))
-		os.Remove(filepath.Join(partdir, "hashes.invalid"))
-		os.Remove(partdir)
-		return
-	}
-	for i := len(suffixDirs) - 1; i > 0; i-- { // shuffle suffixDirs list
-		j := rand.Intn(i + 1)
-		suffixDirs[j], suffixDirs[i] = suffixDirs[i], suffixDirs[j]
-	}
-	for _, suffDir := range suffixDirs {
-		if !needSuffix(filepath.Base(suffDir)) {
-			continue
-		}
-		hashDirs, err := filepath.Glob(filepath.Join(suffDir, "????????????????????????????????"))
-		if err != nil {
-			rd.r.logger.Error("[listObjFiles]", zap.Error(err))
-			return
-		}
-		if len(hashDirs) == 0 {
-			os.Remove(suffDir)
-			continue
-		}
-		for _, hashDir := range hashDirs {
-			fileList, err := filepath.Glob(filepath.Join(hashDir, "*.[tdm]*"))
-			if len(fileList) == 0 {
-				os.Remove(hashDir)
-				continue
-			}
-			if err != nil {
-				rd.r.logger.Error("[listObjFiles]", zap.Error(err))
-				return
-			}
-			for _, objFile := range fileList {
-				select {
-				case objChan <- objFile:
-				case <-cancel:
-					return
-				}
-			}
-		}
-	}
-}
-
-type syncFileArg struct {
-	conn RepConn
-	dev  *ring.Device
-}
-
-func (rd *replicationDevice) syncFile(objFile string, dst []*syncFileArg, handoff bool) (syncs int, insync int, err error) {
-	// TODO: parallelize the data transfer someday
-	var wrs []*syncFileArg
-	lst := strings.Split(objFile, string(os.PathSeparator))
-	relPath := filepath.Join(lst[len(lst)-5:]...)
-	fp, xattrs, fileSize, err := getFile(objFile)
-	if _, ok := err.(quarantineFileError); ok {
-		hashDir := filepath.Dir(objFile)
-		rd.r.logger.Error("[syncFile] Failed audit and is being quarantined",
-			zap.String("hashDir", hashDir),
-			zap.Error(err))
-		QuarantineHash(hashDir)
-		return 0, 0, nil
-	} else if err != nil {
-		return 0, 0, nil
-	}
-	defer fp.Close()
-
-	// are we already going to sync to this region?
-	syncingRemoteRegion := make(map[int]bool)
-
-	// ask each server if we need to sync the file
-	for _, sfa := range dst {
-		var sfr SyncFileResponse
-		thisPath := filepath.Join(sfa.dev.Device, relPath)
-		sfa.conn.SendMessage(SyncFileRequest{Path: thisPath, Xattrs: hex.EncodeToString(xattrs), Size: fileSize,
-			// if we're already syncing handoffs to this remote region, just do a check
-			Check: handoff && syncingRemoteRegion[sfa.dev.Region],
-			// If we're not syncing handoffs, we don't care about the state. Just ping to keep the connection alive.
-			Ping: !handoff && syncingRemoteRegion[sfa.dev.Region],
-		})
-		if err := sfa.conn.RecvMessage(&sfr); err != nil {
-			continue
-		} else if sfr.GoAhead {
-			wrs = append(wrs, sfa)
-			if sfa.dev.Region != rd.dev.Region {
-				syncingRemoteRegion[sfa.dev.Region] = true
-			}
-		} else if sfr.NewerExists {
-			insync++
-			if os.Remove(objFile) == nil {
-				InvalidateHash(filepath.Dir(objFile))
-			}
-		} else if sfr.Exists {
-			insync++
-		}
-	}
-	if len(wrs) == 0 { // nobody needed the file
-		return
-	}
-
-	// send the file to servers
-	scratch := make([]byte, 32768)
-	var length int
-	var totalRead int64
-	for length, err = fp.Read(scratch); err == nil; length, err = fp.Read(scratch) {
-		totalRead += int64(length)
-		for index, sfa := range wrs {
-			if sfa == nil {
-				continue
-			}
-			if _, err := sfa.conn.Write(scratch[0:length]); err != nil {
-				rd.r.logger.Error("Failed to write to remoteDevice",
-					zap.Int("device id", sfa.dev.Id),
-					zap.Error(err))
-				wrs[index] = nil
-			}
-		}
-	}
-	if totalRead != fileSize {
-		return 0, 0, fmt.Errorf("Failed to read the full file: %s, %v", objFile, err)
-	}
-
-	// get file upload results
-	for _, sfa := range wrs {
-		if sfa == nil {
-			continue
-		}
-		var fur FileUploadResponse
-		sfa.conn.Flush()
-		if sfa.conn.RecvMessage(&fur) == nil {
-			if fur.Success {
-				syncs++
-				insync++
-				rd.UpdateStat("FilesSent", 1)
-				rd.UpdateStat("BytesSent", fileSize)
-			}
-		}
-	}
-	return syncs, insync, nil
-}
-
-func (rd *replicationDevice) beginReplication(dev *ring.Device, partition string, hashes bool, rChan chan beginReplicationResponse, headers map[string]string) {
-	var brr BeginReplicationResponse
-	if headers == nil {
-		headers = map[string]string{}
-	}
-	headers["X-Trans-Id"] = fmt.Sprintf("%s-%d", common.UUID(), dev.Id)
-
-	if rc, err := NewRepConn(dev, partition, rd.policy, headers, rd.r.CertFile, rd.r.KeyFile, rd.r.rcTimeout); err != nil {
-		rChan <- beginReplicationResponse{dev: dev, err: err}
-	} else if err := rc.SendMessage(BeginReplicationRequest{Device: dev.Device, Partition: partition, NeedHashes: hashes}); err != nil {
-		rChan <- beginReplicationResponse{dev: dev, err: err}
-	} else if err := rc.RecvMessage(&brr); err != nil {
-		rChan <- beginReplicationResponse{dev: dev, err: err}
-	} else {
-		rChan <- beginReplicationResponse{dev: dev, conn: rc, hashes: brr.Hashes}
-	}
-}
-
-func (rd *replicationDevice) replicateUsingHashes(rjob replJob, moreNodes ring.MoreNodes) {
-	path := filepath.Join(rd.r.deviceRoot, rd.dev.Device, PolicyDir(rd.policy), rjob.partition)
-	syncCount := 0
-	startGetHashesRemote := time.Now()
-	remoteHashes := make(map[int]map[string]string)
-	remoteConnections := make(map[int]RepConn)
-	rChan := make(chan beginReplicationResponse)
-	for _, dev := range rjob.nodes {
-		go rd.i.beginReplication(dev, rjob.partition, true, rChan, rjob.headers)
-	}
-	for i := 0; i < len(rjob.nodes); i++ {
-		rData := <-rChan
-		if rData.err == nil {
-			defer rData.conn.Close()
-			remoteHashes[rData.dev.Id] = rData.hashes
-			remoteConnections[rData.dev.Id] = rData.conn
-		} else if rData.err == RepUnmountedError {
-			if nextNode := moreNodes.Next(); nextNode != nil {
-				go rd.i.beginReplication(nextNode, rjob.partition, true, rChan, rjob.headers)
-				rjob.nodes = append(rjob.nodes, nextNode)
-			} else {
-				break
-			}
-		}
-	}
-	if len(remoteHashes) == 0 {
-		return
-	}
-
-	timeGetHashesRemote := float64(time.Now().Sub(startGetHashesRemote)) / float64(time.Second)
-	startGetHashesLocal := time.Now()
-
-	recalc := []string{}
-	hashes, err := GetHashes(rd.r.deviceRoot, rd.dev.Device, rjob.partition, recalc, rd.r.reclaimAge, rd.policy, rd.r.logger)
-	if err != nil {
-		rd.r.logger.Error("[replicateUsingHashes] error getting local hashes", zap.Error(err))
-		return
-	}
-	for suffix, localHash := range hashes {
-		for _, remoteHash := range remoteHashes {
-			if remoteHash[suffix] != "" && localHash != remoteHash[suffix] {
-				recalc = append(recalc, suffix)
-				break
-			}
-		}
-	}
-	hashes, err = GetHashes(rd.r.deviceRoot, rd.dev.Device, rjob.partition, recalc, rd.r.reclaimAge, rd.policy, rd.r.logger)
-	if err != nil {
-		rd.r.logger.Error("[replicateUsingHashes] error recalculating local hashes", zap.Error(err))
-		return
-	}
-	timeGetHashesLocal := float64(time.Now().Sub(startGetHashesLocal)) / float64(time.Second)
-
-	objChan := make(chan string, 100)
-	cancel := make(chan struct{})
-	defer close(cancel)
-	go rd.i.listObjFiles(objChan, cancel, path, func(suffix string) bool {
-		for _, remoteHash := range remoteHashes {
-			if hashes[suffix] != remoteHash[suffix] {
-				return true
-			}
-		}
-		return false
-	})
-	startSyncing := time.Now()
-	for objFile := range objChan {
-		toSync := make([]*syncFileArg, 0)
-		suffix := filepath.Base(filepath.Dir(filepath.Dir(objFile)))
-		for _, dev := range rjob.nodes {
-			if rhashes, ok := remoteHashes[dev.Id]; ok && hashes[suffix] != rhashes[suffix] {
-				if !remoteConnections[dev.Id].Disconnected() {
-					toSync = append(toSync, &syncFileArg{conn: remoteConnections[dev.Id], dev: dev})
-				}
-			}
-		}
-		if len(toSync) == 0 {
-			break
-		}
-		if syncs, _, err := rd.i.syncFile(objFile, toSync, false); err == nil {
-			syncCount += syncs
-		} else {
-			rd.r.logger.Error("[syncFile]", zap.Error(err))
-			return
-		}
-	}
-	for _, conn := range remoteConnections {
-		if !conn.Disconnected() {
-			conn.SendMessage(SyncFileRequest{Done: true})
-		}
-	}
-	timeSyncing := float64(time.Now().Sub(startSyncing)) / float64(time.Second)
-	if syncCount > 0 {
-		rd.r.logger.Info("[replicateUsingHashes]",
-			zap.String("Partition", path),
-			zap.Any("Files Synced", syncCount),
-			zap.Float64("timeGetHashesRemote", timeGetHashesRemote),
-			zap.Float64("timeGetHashesLocal", timeGetHashesLocal),
-			zap.Float64("timeSyncing", timeSyncing))
-	}
-}
-
-func (rd *replicationDevice) replicateAll(rjob replJob, isHandoff bool) {
-	path := filepath.Join(rd.r.deviceRoot, rd.dev.Device, PolicyDir(rd.policy), rjob.partition)
-	syncCount := 0
-	remoteConnections := make(map[int]RepConn)
-	rChan := make(chan beginReplicationResponse)
-	for _, dev := range rjob.nodes {
-		go rd.i.beginReplication(dev, rjob.partition, false, rChan, rjob.headers)
-	}
-	for i := 0; i < len(rjob.nodes); i++ {
-		rData := <-rChan
-		if rData.err == nil {
-			defer rData.conn.Close()
-			remoteConnections[rData.dev.Id] = rData.conn
-		}
-	}
-	if len(remoteConnections) == 0 {
-		return
-	}
-
-	objChan := make(chan string, 100)
-	cancel := make(chan struct{})
-	defer close(cancel)
-	go rd.i.listObjFiles(objChan, cancel, path, func(string) bool { return true })
-	for objFile := range objChan {
-		toSync := make([]*syncFileArg, 0)
-		for _, dev := range rjob.nodes {
-			if remoteConnections[dev.Id] != nil && !remoteConnections[dev.Id].Disconnected() {
-				toSync = append(toSync, &syncFileArg{conn: remoteConnections[dev.Id], dev: dev})
-			}
-		}
-		if len(toSync) == 0 {
-			return
-		}
-		if syncs, insync, err := rd.i.syncFile(objFile, toSync, true); err == nil {
-			syncCount += syncs
-
-			success := insync == len(rjob.nodes)
-			if rd.r.quorumDelete {
-				success = insync >= len(rjob.nodes)/2+1
-			}
-			if success && isHandoff {
-				os.Remove(objFile)
-				os.Remove(filepath.Dir(objFile))
-			}
-		} else {
-			rd.r.logger.Error("[syncFile]", zap.Error(err))
-			return
-		}
-	}
-	for _, conn := range remoteConnections {
-		if !conn.Disconnected() {
-			conn.SendMessage(SyncFileRequest{Done: true})
-		}
-	}
-	if syncCount > 0 {
-		rd.r.logger.Info("[replicateAll]", zap.String("Partition", path), zap.Any("Files Synced", syncCount))
-	}
-}
-
-func (rd *replicationDevice) Key() string {
-	return deviceKeyId(rd.dev.Device, rd.policy)
-}
-
-func (rd *replicationDevice) cleanTemp() {
-	tempDir := TempDirPath(rd.r.deviceRoot, rd.dev.Device)
-	if tmpContents, err := ioutil.ReadDir(tempDir); err == nil {
-		for _, tmpEntry := range tmpContents {
-			if time.Since(tmpEntry.ModTime()) > tmpEmptyTime {
-				os.RemoveAll(filepath.Join(tempDir, tmpEntry.Name()))
-			}
-		}
-	}
-}
-
-func (rd *replicationDevice) replicatePartition(partition string) {
-	rd.r.replicateConcurrencySem <- struct{}{}
-	defer func() {
-		<-rd.r.replicateConcurrencySem
-	}()
-	partitioni, err := strconv.ParseUint(partition, 10, 64)
-	if err != nil {
-		return
-	}
-	nodes, handoff := rd.r.objectRings[rd.policy].GetJobNodes(partitioni, rd.dev.Id)
-	policy := rd.r.policies[rd.policy]
-	if policy == nil {
-		return
-	}
-	rjob := replJob{partition: partition, nodes: nodes}
-	if handoff || (policy.Type == "replication-nursery" &&
-		!common.LooksTrue(policy.Config["cache_hash_dirs"])) {
-		rd.i.replicateAll(rjob, handoff)
-	} else {
-		rd.i.replicateUsingHashes(rjob, rd.r.objectRings[rd.policy].GetMoreNodes(partitioni))
-	}
-	rd.UpdateStat("PartitionsDone", 1)
-}
-
-func (rd *replicationDevice) listPartitions() ([]string, []string, error) {
-	// returns a list of all partitions and a subset of that list- just the handoffs
-	objPath := filepath.Join(rd.r.deviceRoot, rd.dev.Device, PolicyDir(rd.policy))
-	partitions, err := filepath.Glob(filepath.Join(objPath, "[0-9]*"))
-	if err != nil {
-		return nil, nil, err
-	}
-	partitionList := make([]string, 0, len(partitions))
-	handoffList := []string{}
-	for _, partition := range partitions {
-		partition = filepath.Base(partition)
-		if len(rd.r.partitions) > 0 && !rd.r.partitions[partition] {
-			continue
-		}
-		if _, err := strconv.ParseUint(partition, 10, 64); err == nil {
-			partitionList = append(partitionList, partition)
-		}
-	}
-	for i := len(partitionList) - 1; i > 0; i-- { // shuffle partition list
-		j := rand.Intn(i + 1)
-		partitionList[j], partitionList[i] = partitionList[i], partitionList[j]
-	}
-	for _, partition := range partitionList {
-		if pi, err := strconv.ParseUint(partition, 10, 64); err == nil {
-			if _, handoff := rd.r.objectRings[rd.policy].GetJobNodes(pi, rd.dev.Id); handoff {
-				handoffList = append(handoffList, partition)
-			}
-		}
-	}
-	return partitionList, handoffList, nil
-}
-
-func (rd *replicationDevice) Replicate() {
-	if rd.r.policies[rd.policy].Type == "hec" { // TODO yuck
-		return
-	}
-	defer srv.LogPanics(rd.r.logger, fmt.Sprintf("PANIC REPLICATING DEVICE: %s", rd.dev.Device))
-	rd.UpdateStat("startRun", 1)
-	if mounted, err := fs.IsMount(filepath.Join(rd.r.deviceRoot, rd.dev.Device)); rd.r.checkMounts && (err != nil || mounted != true) {
-		rd.r.logger.Error("[replicateDevice] Drive not mounted", zap.String("Device", rd.dev.Device), zap.Error(err))
-		return
-	}
-	if fs.Exists(filepath.Join(rd.r.deviceRoot, rd.dev.Device, "lock_device")) {
-		return
-	}
-
-	rd.i.cleanTemp()
-
-	allPartitionList, handoffPartitions, err := rd.i.listPartitions()
-	if err != nil {
-		rd.r.logger.Error("[replicateDevice] Error getting partition list",
-			zap.String("Device", rd.dev.Device),
-			zap.Error(err))
-		return
-	} else if len(allPartitionList) == 0 {
-		rd.r.logger.Info("[replicateDevice] No partitions found",
-			zap.String("filepath", filepath.Join(rd.r.deviceRoot, rd.dev.Device, PolicyDir(rd.policy))))
-		return
-	}
-	rd.UpdateStat("PartitionsTotal", int64(len(allPartitionList)))
-
-	lastListing := time.Now()
-	handoffsForLog := len(handoffPartitions)
-	for i, partition := range allPartitionList {
-		rd.UpdateStat("checkin", 1)
-		select {
-		case <-rd.cancel:
-			{
-				rd.r.logger.Error("replicateDevice canceled for device", zap.String("Device", rd.dev.Device))
-				return
-			}
-		default:
-		}
-		rd.processPriorityJobs()
-		rd.i.replicatePartition(partition)
-		if j := common.StringInSliceIndex(partition, handoffPartitions); j >= 0 {
-			handoffPartitions = append(handoffPartitions[:j], handoffPartitions[j+1:]...)
-		}
-		time.Sleep(replicatePartSleepTime)
-		if i%handoffToAllMod == 0 && len(handoffPartitions) > 0 {
-			var p string
-			p, handoffPartitions = handoffPartitions[0], handoffPartitions[1:]
-			rd.i.replicatePartition(p)
-		}
-		if len(handoffPartitions) == 0 {
-			if handoffsForLog > 0 {
-				rd.r.logger.Info("[replicateDevice] Completed handoff replication pass",
-					zap.Int("handoffsProcessed", handoffsForLog),
-					zap.Duration("handoffDuration", time.Since(lastListing)))
-				handoffsForLog = 0
-			}
-			if time.Since(lastListing) > handoffListDirFreq {
-				if _, handoffPartitions, err = rd.i.listPartitions(); err != nil {
-					rd.r.logger.Error("[replicateDevice] Error getting handoff partition list",
-						zap.String("Device", rd.dev.Device),
-						zap.Error(err))
-				}
-				lastListing = time.Now()
-				handoffsForLog = len(handoffPartitions)
-			}
-		}
-	}
-	rd.UpdateStat("FullReplicateCount", 1)
-}
-
-func (rd *replicationDevice) Cancel() {
-	rd.UpdateStat("cancel", 1)
-	close(rd.cancel)
-}
-
-func (rd *replicationDevice) ReplicateLoop() {
-	for {
-		select {
-		case <-rd.cancel:
-			return
-		default:
-			rd.Replicate()
-		}
-		time.Sleep(replicateLoopSleepTime)
-	}
-}
-
-type NoMoreNodes struct{}
-
-func (n *NoMoreNodes) Next() *ring.Device {
-	return nil
-}
-
-func (rd *replicationDevice) PriorityReplicate(w http.ResponseWriter, pri PriorityRepJob) error {
-	if !fs.Exists(filepath.Join(rd.r.deviceRoot, pri.FromDevice.Device, PolicyDir(rd.policy), strconv.FormatUint(pri.Partition, 10))) {
-		w.WriteHeader(404)
-		return nil
-	}
-	timer := time.NewTimer(priorityReplicateTimeout)
-	defer timer.Stop()
-	select {
-	case rd.priRep <- pri:
-		w.WriteHeader(200)
-		return nil
-	case <-timer.C:
-		return fmt.Errorf("timed out waiting for chan")
-	}
-}
-
-// processPriorityJobs runs any pending priority jobs given the device's id
-func (rd *replicationDevice) processPriorityJobs() {
-	for {
-		select {
-		case pri := <-rd.priRep:
-			func() {
-				time.Sleep(replicatePartSleepTime)
-				rd.r.replicateConcurrencySem <- struct{}{}
-				defer func() {
-					<-rd.r.replicateConcurrencySem
-				}()
-				partition := strconv.FormatUint(pri.Partition, 10)
-				_, handoff := rd.r.objectRings[rd.policy].GetJobNodes(pri.Partition, pri.FromDevice.Id)
-				jobType := "local"
-				if handoff {
-					jobType = "handoff"
-				}
-				policy := rd.r.policies[rd.policy]
-				if policy == nil {
-					return
-				}
-				rd.r.logger.Info("PriorityReplicationJob",
-					zap.Uint64("partition", pri.Partition),
-					zap.String("jobType", jobType),
-					zap.String("From Device", pri.FromDevice.Device),
-					zap.String("To Device", pri.ToDevice.Device))
-				rjob := replJob{
-					partition: partition, nodes: []*ring.Device{pri.ToDevice},
-					headers: map[string]string{"X-Force-Acquire": "true"}}
-				if handoff || (policy.Type == "replication-nursery" &&
-					!common.LooksTrue(policy.Config["cache_hash_dirs"])) {
-					rd.i.replicateAll(rjob, handoff)
-				} else {
-					rd.i.replicateUsingHashes(rjob, &NoMoreNodes{})
-				}
-			}()
-			rd.UpdateStat("PriorityRepsDone", 1)
-		default:
-			return
-		}
-	}
 }
 
 // Object replicator daemon object
